Fixes event loops in all code files
The event loop used window.isOpen() as the loop condition, and closed the window too early when SFML indicated the window should close.
The loop condition is now based on a boolean flag. The window is closed at the end of the program, after all other OpenGL calls have been made.