Fix spacing issue
On the website, in the 'device coordinates' section, if your screen is narrow enough, you can see that the two images appear to be one image, since they're not separated. This isn't something that happens in the Github preview, but it does happen on the website.

This PR adds some padding to the bottom of the first image to make it obvious that the images are separate.